@@ -33,43 +33,38 @@ namespace Digipolis.Web.SampleApi.Controllers
         /// <remarks>
         /// This endpoint demostrates the use of the old <see cref="PagedResult{T}"/> and it's resulting json.
         /// </remarks>
-        /// <param name="queryOptions">Query options from uri</param>
         /// <returns>An array of value objects</returns>
-        [HttpGet()]
+        [HttpGet]
         [ProducesResponseType(typeof(PagedResult<ValueDto>), 200)]
         [AllowAnonymous]
-        [Versions(Versions.V1, Versions.V2)]
         [Produces("application/hal+json")]
+        [Versions(Versions.V1, Versions.V2)]
         public IActionResult Get([FromQuery]CriteriaDto criteria)
         {
-            int total;
-            var values = _valueLogic.GetAll(criteria, out total);
-            //var result = queryOptions.ToPagedResult(values, total, "kevin", new { test = 0 });
-            var result = criteria.ToPagedResult(values, total, "Get", "Values", new { test = 0 });
+            var values = _valueLogic.GetAll(criteria, out var total);
+            var result = criteria.ToPagedResult(values, total, ControllerContext.ActionDescriptor.ActionName, ControllerContext.ActionDescriptor.ControllerName);
             return Ok(result);
         }
-
+        
         /// <summary>
         /// Get all values 
         /// </summary>
         /// <remarks>
         /// This endpoint demostrates the use of the new <see cref="PagedResult{T, EmbeddedT}"/> and it's resulting json.
         /// </remarks>
-        /// <param name="queryOptions">Query options from uri</param>
+        /// <param name="criteria">Query options from uri</param>
         /// <returns>An array of value objects</returns>
         [HttpGet("new")]
-        [ProducesResponseType(typeof(PagedResult<ValueDto, EmbeddedValueDto>), 200)]
         [AllowAnonymous]
         [Versions(Versions.V1, Versions.V2)]
         [Produces("application/hal+json")]
         public IActionResult GetNew([FromQuery]CriteriaDto criteria)
         {
-            int total;
-            var values = _valueLogic.GetAll(criteria, out total);
-            //var result = queryOptions.ToPagedResult(values, total, "kevin", new { test = 0 });
+            var values = _valueLogic.GetAll(criteria, out var total);
             var result = criteria.ToPagedResult<ValueDto, EmbeddedValueDto>(values, total);
             return Ok(result);
         }
+        
         /// <summary>
         /// Get a value by id
         /// </summary>
@@ -79,13 +74,14 @@ namespace Digipolis.Web.SampleApi.Controllers
         [ProducesResponseType(typeof(ValueDto), 200)]
         [ProducesResponseType(typeof(ValueDto), 401)]
         [AllowAnonymous]
+        [Versions(Versions.V1, Versions.V2)]
         [Produces("application/json", "text/csv")]
         public IActionResult Get(int id)
         {
             var value = _valueLogic.GetById(id);
             return Ok(value);
         }
-
+        
         /// <summary>
         /// Add a new value
         /// </summary>
@@ -101,7 +97,7 @@ namespace Digipolis.Web.SampleApi.Controllers
             value = _valueLogic.Add(value);
             return CreatedAtAction("Get", new { id = value.Id }, value);
         }
-
+        
         /// <summary>
         /// Update an existing value object
         /// </summary>
@@ -117,7 +113,7 @@ namespace Digipolis.Web.SampleApi.Controllers
             value = _valueLogic.Update(id, value);
             return Ok(value);
         }
-
+        
         /// <summary>
         /// Delete a value by it's Id
         /// </summary>
@@ -130,9 +126,9 @@ namespace Digipolis.Web.SampleApi.Controllers
             _valueLogic.Delete(id);
             return NoContent();
         }
-
+        
         /// <summary>
-        /// Thorws an exception
+        /// Throws an exception
         /// </summary>
         /// <returns></returns>
         [HttpGet("exception")]
